# uni-api_lovpen.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "uni-api",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/files/upload": {
      "post": {
        "summary": "Upload File",
        "deprecated": false,
        "description": "上传文件",
        "operationId": "upload_file_files_upload_post",
        "tags": [
          "User Files"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "title": "File",
                    "example": ""
                  },
                  "parent_id": {
                    "type": "string",
                    "title": "Parent Id",
                    "default": "",
                    "example": ""
                  },
                  "description": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "title": "Description",
                    "example": ""
                  },
                  "is_public": {
                    "type": "boolean",
                    "title": "Is Public",
                    "default": false,
                    "example": "false"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/files/": {
      "get": {
        "summary": "List Files",
        "deprecated": false,
        "description": "列出文件",
        "operationId": "list_files_files__get",
        "tags": [
          "User Files"
        ],
        "parameters": [
          {
            "name": "parent_id",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFileListResponse"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/files/{file_id}": {
      "get": {
        "summary": "Get File",
        "deprecated": false,
        "description": "获取文件信息",
        "operationId": "get_file_files__file_id__get",
        "tags": [
          "User Files"
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFileResponse"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "Update File",
        "deprecated": false,
        "description": "更新文件信息",
        "operationId": "update_file_files__file_id__put",
        "tags": [
          "User Files"
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserFileUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserFileResponse"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "Delete File",
        "deprecated": false,
        "description": "删除文件或目录",
        "operationId": "delete_file_files__file_id__delete",
        "tags": [
          "User Files"
        ],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileOperationResponse"
                }
              }
            },
            "headers": {}
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FileOperationResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "file_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "FileOperationResponse",
        "description": "文件操作响应"
      },
      "FileUploadResponse": {
        "properties": {
          "file_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          },
          "file_url": {
            "type": "string",
            "title": "File Url"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "file_size",
          "file_type",
          "file_url",
          "success",
          "message"
        ],
        "title": "FileUploadResponse",
        "description": "文件上传响应"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserFileListResponse": {
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/UserFileResponse"
            },
            "type": "array",
            "title": "Files"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "files",
          "total",
          "page",
          "page_size",
          "total_pages"
        ],
        "title": "UserFileListResponse",
        "description": "文件列表响应"
      },
      "UserFileResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "title": "Filename",
            "description": "原始文件名"
          },
          "file_path": {
            "type": "string",
            "maxLength": 500,
            "title": "File Path",
            "description": "存储路径"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "description": "文件大小（字节）"
          },
          "file_type": {
            "type": "string",
            "maxLength": 100,
            "title": "File Type",
            "description": "文件类型/MIME类型"
          },
          "is_directory": {
            "type": "boolean",
            "title": "Is Directory",
            "description": "是否为目录",
            "default": false
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id",
            "description": "父目录ID"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "文件描述"
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "description": "是否公开",
            "default": false
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "file_path",
          "file_size",
          "file_type",
          "id",
          "user_id",
          "created_at",
          "updated_at"
        ],
        "title": "UserFileResponse",
        "description": "文件响应"
      },
      "UserFileUpdate": {
        "properties": {
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_public": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Public"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          }
        },
        "type": "object",
        "title": "UserFileUpdate",
        "description": "更新文件请求"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}